# FIX: Sparse matrix addition/subtraction eliminates explicit zeros