Fixes #9318 
Binary operations on Sparse matrices removes the explicit zeros. These changes preserve the explicit zeros in the output matrix.