@perimosocordiae 2019-01-02
Can you run an informal benchmark to show the impact on running time / memory usage? It seems like there shouldn't be too much impact in the no-explicit-zeros case, but I'd like to have some numbers.