@ananyashreyjain 2019-01-02
Benchmark reveals that there is some impact on the running time.

   | before       |after              |  ratio  |                                                                              |
   |---------------|------------------|---------|---------------------------------------------------------------------------|
   | [a2ffe09a]| 20b3c18c]   |           |                                                                                           |
   | 1.78ms      | 2.27ms       |  1 .28 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__sub__')|
   |  4.56ms     | 5.76ms       | 1.26   | sparse.Arithmetic.time_arithmetic('csr', 'AB', 'multiply') |
   | 4.45ms      | 5.50ms       | 1.23   | sparse.Arithmetic.time_arithmetic('csr', 'BB', '__add__')|
   | 1.89ms      | 2.33ms       | 1.23   | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__add__') |
   | 3.97ms      | 4.85ms       | 1.22   | sparse.Arithmetic.time_arithmetic('csr', 'BA', '__add__') |
   | 4.54ms      | 5.54ms       | 1.22   | sparse.Arithmetic.time_arithmetic('csr', 'BB', 'multiply') |
   | 3.98ms      | 4.85ms       | 1.22   | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__sub__') |
   | 4.00ms      | 4.88ms       | 1.22   | sparse.Arithmetic.time_arithmetic('csr', 'BA', '__sub__') |
   | 3.96ms      | 4.82ms       | 1.22   | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__add__') |
   | 4.57ms      | 5.53ms       | 1.21   | sparse.Arithmetic.time_arithmetic('csr', 'BA', 'multiply') |
   | 2.01ms      | 2.37ms       | 1.18   | sparse.Arithmetic.time_arithmetic('csr', 'AA', 'multiply') |
   | 4.66ms      | 5.49ms       | 1.18   | sparse.Arithmetic.time_arithmetic('csr', 'BB', '__sub__') |
   | 33.01ms    | 35.15ms     | 1.06   | sparse.Arithmetic.time_arithmetic('csr', 'BB', '__mul__') |