@perimosocordiae 2019-01-09
My main concern is a performance hit for matrices without explicit zeros, which I expect is the most common case, and isn't impacted by the issue this PR fixes.

One approach might be to maintain two separate functions: the current one that ignores zeros, and a new one that preserves them. You could dispatch to the appropriate function based on known properties of the matrices, or even try checking for explicit zeros before dispatching.

That said, it may also be possible to improve the algorithm for preserving explicit zeros to make the overhead negligible in the no-zeros case, but I haven't looked into it very deeply.