@ananyashreyjain 2019-01-14
Benchmark for the new commit : 

| before        |   after           |ratio|                                                                                           |
|----------------|-----------------|-------|---------------------------------------------------------------------------|
|  [a2ffe09a]  | [750fe675]  |        |                                                                                          |
|    1.91ms    |  2.17ms      |1.13 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__add__')|
|    1.97ms    | 2.19ms       |1.11 | sparse.Arithmetic.time_arithmetic('csr', 'AA', 'multiply')|
|    1.79ms    | 1.98ms      |1.10  |sparse.Arithmetic.time_arithmetic('csr', 'AA', '__sub__')|
|    4.03ms    | 4.36ms      |1.08  |sparse.Arithmetic.time_arithmetic('csr', 'BA', '__sub__')|
|    4.53ms    | 4.87ms      |1.07  |sparse.Arithmetic.time_arithmetic('csr', 'BB', '__add__')|
|    4.57ms    |4.86ms      |1.06  |sparse.Arithmetic.time_arithmetic('csr', 'BA', 'multiply')|
|   14.77ms   |15.63ms     | 1.06 |sparse.Arithmetic.time_arithmetic('csr', 'AB', '__mul__')|
|    4.11ms    |4.33ms      |1.05  |sparse.Arithmetic.time_arithmetic('csr', 'BA', '__add__')|
|    4.71ms    |4.17ms      |0.89  |sparse.Arithmetic.time_arithmetic('csr', 'AB', 'multiply')|
|    5.59ms    | 4.89ms      |0.88 | sparse.Arithmetic.time_arithmetic('csr', 'BB', 'multiply')|
|   44.59ms   | 35.17ms     | 0.79  |sparse.Arithmetic.time_arithmetic('csr', 'BB', '__mul__')|