@ananyashreyjain 2019-01-14
@perimosocordiae  `csr_binop_csr_canonical()` is designed to not include explicit zeros so it checks the uncommon non-zero elements of both matrices so as to not include them in the final matrix . I removed those checks which may be the cause of benchmark results getting better for some cases.