@ananyashreyjain 2019-01-14
Benchmark :

|    before   |  after             |ratio|                                                                                           |
|--------------|------------------|-------|---------------------------------------------------------------------------|
| [a2ffe09a]|  [6929a0ff]    |        |                                                                                          |
|    1.93ms |    2.05ms      |1.06 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__sub__')|
|    2.09ms |    2.20ms      |1.05  |sparse.Arithmetic.time_arithmetic('csr', 'AA', '__add__')|
|    4.50ms |    4.27ms      |0.95  |sparse.Arithmetic.time_arithmetic('csr', 'BA', '__sub__')|
|    5.12ms |    4.79ms      |0.94  |sparse.Arithmetic.time_arithmetic('csr', 'BB', '__add__')|
|    5.37ms |    4.89ms      |0.91  |sparse.Arithmetic.time_arithmetic('csr', 'AB', 'multiply')  |
|   18.55ms|    16.84ms    | 0.91  |sparse.Arithmetic.time_arithmetic('csr', 'BA', '__mul__')|
|  17.31ms |   15.34ms     | 0.89 | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__mul__')|