@ananyashreyjain 2019-01-20
Benchmark:
 | before    | after       |ratio|                                                                                                |
|--------------|-------------|------|-------------------------------------------------------------------------------|
  |[95029ad1]| [24433a94]|                                                                                                |
  | 2.01ms     |2.27ms      |1.13  sparse.Arithmetic.time_arithmetic('csr', 'AA', '__add__')|
   | 2.05ms     |2.29ms      |1.12  sparse.Arithmetic.time_arithmetic('csr', 'AA', 'multiply')|
   | 1.85ms     |2.03ms      |1.10  sparse.Arithmetic.time_arithmetic('csr', 'AA', '__sub__')|
   |4.18ms     |4.53ms      |1.08  sparse.Arithmetic.time_arithmetic('csr', 'BA', '__sub__')|
   |15.22ms    |16.11ms    |  1.06  sparse.Arithmetic.time_arithmetic('csr', 'AB', '__mul__')|
    |5.31ms     |5.02ms      |0.95  sparse.Arithmetic.time_arithmetic('csr', 'BB', 'multiply')|
   |38.28ms    |35.40ms     | 0.92  sparse.Arithmetic.time_arithmetic('csr', 'BB', '__mul__')|
    |4.78ms     |4.34ms     | 0.91  sparse.Arithmetic.time_arithmetic('csr', 'AB', 'multiply')|

@perimosocordiae reminder for review.