@ananyashreyjain 2019-01-22
Initially `result` was  added to `Cx` if it was not zero because we don't want zero values in `Cx` but this also removed the explicitly defined zero values . To include those zeros we need to check  `Ax` and `Bx` for zeros (Ax and Bx can have zeros only when they are explicitly defined by user) so in the case when matrix B has no non-zero values at the indices for which A has some non-zero values then the values in A  at those indices were included in the final matrix if they were not zero but Ax cannot have zero values unless they are explicitly defines by user , so I removed the condition which was checking `Ax` for zeros. Same goes for `Bx`. In the case when there is a non-zero value in both the matrices  at some index and value of `result` is zero we only need to check  the value in one matrix because if value in one matrix is zero then value in other matrix  at that index should be zero in case of addition and subtraction. 
I extended similar logic for `csr_binop_csr_general()` but I couldn't fully understand its working so I am little doubtful for the non-canonical matrices.